Item deletion

+++ app/controllers/items_controller.rb
@@ -19,8 +19,13 @@ class ItemsController<ApplicationController
 
   def create
     merchant = Merchant.find(params[:merchant_id])
-    merchant.items.create(item_params)
-    redirect_to "/merchants/#{merchant.id}/items"
+    item = merchant.items.create(item_params)
+    if item.save
+      redirect_to "/merchants/#{merchant.id}/items"
+    else
+      flash.notice = item.errors.full_messages.to_sentence
+      redirect_to "/merchants/#{merchant.id}/items/new"
+    end
   end
 
   def edit
@@ -30,13 +35,29 @@ class ItemsController<ApplicationController
   def update
     item = Item.find(params[:id])
     item.update(item_params)
-    redirect_to "/items/#{item.id}"
+    if item.save
+      redirect_to "/items/#{item.id}"
+    else
+      flash.notice = item.errors.full_messages.to_sentence
+      redirect_to "/items/#{item.id}/edit"
+    end
   end
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to "/items"
+    item_present = ItemOrder.pluck(:item_id).include?(item.id)
+    if item_present
+      flash.notice = 'Cannot delete, this item has orders in progress.'
+      redirect_to "/items/#{params[:id]}"
+    else
+      if cart.contents.has_key?(item.id.to_s)
+        cart.contents.delete(item.id.to_s)
+        session[:cart] = cart.contents
+      end
+      Review.delete(Review.where("item_id = #{item.id}"))
+      item.destroy
+      redirect_to "/items"
+    end
   end
 
   private
